Refactor afterglow_models into organized submodule structure
- Move afterglow_models.py → afterglow_models/base_models.py
- Update __init__.py to import from base_models.py and vegas_models.py
- Explicitly export private functions (_get_kn_dynamics, _pnu_synchrotron)
- Maintains full backward compatibility
- All existing imports continue to work
- Tests pass

Structure:
  afterglow_models/
    __init__.py         - Unified imports
    base_models.py      - Original afterglow models (2424 lines)
    vegas_models.py     - VegasAfterglow models (550 lines)

No breaking changes - all imports from afterglow_models work as before.
