+++ redback/transient_models/afterglow_models/base_models.py
